requirements.txt
Django==4.1.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
django-filter==22.1
Pillow==9.4.0